+++ src/config/passport.ts
@@ -1,8 +1,9 @@
 import passport from 'passport';
 import { Strategy as JwtStrategy, ExtractJwt, StrategyOptionsWithRequest, VerifiedCallback } from 'passport-jwt';
-import { Strategy as GoogleStrategy } from 'passport-google-oauth20';
 import { PrismaClient } from '@prisma/client';
 import { Request } from 'express';
+import { configureGoogleStrategy } from './social/google';
+import { configureNaverStrategy } from './social/naver';
 
 const prisma = new PrismaClient();
 
@@ -14,6 +15,7 @@ const options: StrategyOptionsWithRequest = {
 };
 
 passport.use(
+  'jwt',
   new JwtStrategy(
     options,
     async (req: Request, jwtPayload: any, done: VerifiedCallback) => {
@@ -33,59 +35,14 @@ passport.use(
   )
 );
 
-// Google OAuth Strategy 설정
-passport.use(
-  new GoogleStrategy(
-    {
-      clientID: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-      callbackURL: process.env.GOOGLE_CALLBACK_URL!,
-    },
-    async (accessToken, refreshToken, profile, done) => {
-      try {
-        // Google 프로필에서 필요한 정보 추출
-        const googleId = profile.id;
-        const email = profile.emails?.[0]?.value;
-        const name = profile.displayName;
-        const picture = profile.photos?.[0]?.value;
+// 소셜 로그인 Strategy 설정
+configureGoogleStrategy();
+configureNaverStrategy();
 
-        if (!email) {
-          return done(new Error('Google 계정에서 이메일을 가져올 수 없습니다.'), false);
-        }
-
-        // 기존 사용자 확인
-        let user = await prisma.user.findFirst({
-          where: {
-            email: email,
-            social_type: 'google'
-          }
-        });
-
-        if (!user) {
-          // 새 사용자 생성
-          user = await prisma.user.create({
-            data: {
-              email: email,
-              name: name || 'Unknown',
-              nickname: name || 'Unknown',
-              social_type: 'google',
-              status: true,
-              role: 'USER'
-            }
-          });
-        }
-
-        return done(null, user);
-      } catch (error) {
-        return done(error as Error, false);
-      }
-    }
-  )
-);
 
 // Passport serialize/deserialize 설정
 passport.serializeUser((user: any, done) => {
-  done(null, user.user_id); 
+  done(null, user.user_id);
 });
 
 passport.deserializeUser(async (id: number, done) => {
@@ -101,4 +58,5 @@ passport.deserializeUser(async (id: number, done) => {
 
 export const authenticateJwt = passport.authenticate('jwt', { session: false });
 export const authenticateGoogle = passport.authenticate('google', { scope: ['profile', 'email'] });
+export const authenticateNaver = passport.authenticate('naver', { authType: 'reprompt' });
 export default passport;
