+++ src/routes/auth.ts
@@ -1,41 +1,14 @@
 import { Router, Request, Response } from 'express';
-import passport from 'passport';
-import jwt from 'jsonwebtoken';
 import { PrismaClient } from '@prisma/client';
+import googleAuthRouter from './auth/google';
+import naverAuthRouter from './auth/naver';
 
 const router = Router();
 const prisma = new PrismaClient();
 
-// Google OAuth 로그인 시작
-router.get('/google', passport.authenticate('google', { scope: ['profile', 'email'] }));
-
-// Google OAuth 콜백 처리
-router.get('/google/callback', 
-  passport.authenticate('google', { failureRedirect: '/login' }),
-  async (req: Request, res: Response) => {
-    try {
-      const user = req.user as any;
-      
-      if (!user) {
-        res.status(401).json({ message: '인증에 실패했습니다.' });
-        return;
-      }
-
-      // JWT 토큰 생성 (user_id를 Int로 바로 사용)
-      const token = jwt.sign(
-        { id: user.user_id, email: user.email },
-        process.env.JWT_SECRET!,
-        { expiresIn: '7d' }
-      );
-
-      // 성공 시 리다이렉트 (프론트엔드 URL로 변경 필요)
-      res.redirect(`/auth/success?token=${token}`);
-    } catch (error) {
-      console.error('Google OAuth callback error:', error);
-      res.redirect('/auth/error');
-    }
-  }
-);
+// 소셜 로그인 라우터
+router.use('/google', googleAuthRouter);
+router.use('/naver', naverAuthRouter);
 
 // 로그인 성공 페이지 (토큰 전달용)
 router.get('/success', (req: Request, res: Response) => {
